pyproject.toml
[tool.poetry]
name = "bunnywriting-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Daniil <39118675+TyPaporotnyk@users.noreply.github.com>"]
license = "MIT License"
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.4.1"
sqlalchemy = "^2.0.27"
alembic = "^1.13.1"
aiomysql = "^0.2.0"
pymysql = "^1.1.0"
redis = "^5.0.1"
aioredis = "^2.0.1"
loguru = "^0.7.2"
pydantic-settings = "^2.2.0"
cryptography = "^42.0.3"
celery = "^5.3.6"

[tool.poetry.group.dev]
optional=false

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pytest-asyncio = "^0.23.5"
pytest = "^8.0.1"


[tool.black]
line-length = 119
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 119
known_first_party = [
    "src",
]
skip = ["venv/"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
